Solution to day 9 part 2
The previous direction system, designed for two knots, had a flaw when
used with a bigger rope: a knot that moves diagonally saves its
`lastDirection` as a diagonal vector. If this diagonal motion made the
knot behind it within the same line or column, then it would end up
moving diagonally after using its parent's `lastDirection`. This
situation only occurs with more than three knots.

This was for the better because during the review, I realized a much
better mechanism for deriving the direction: it has to be the closest
vector (in terms of the angle between them) to the vector segment formed
by the two consecutive knots, whose coordinates are either 1 or 0. The
`signum` method does that. Beautiful.